requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
starlette==0.34.0
attrs==23.1.0
requests==2.31.0
python-dotenv==1.0.0
